Rename MdnsServicesListView and related objects
The MdnsServiceListView, as well as the corresponding ListAdapter and fragment,
were used to display WebRTC services as well. So we rename it to something more
generic.
