+++ babble/src/main/java/io/mosaicnetworks/babble/configure/package-info.java
@@ -23,7 +23,7 @@
  */
 
 /**
- * io.mosaicnetworks.babble.configure contains UI components to display discovered babble instances and
- * fragments to join / create new groups
+ * io.mosaicnetworks.babble.configure contains UI components to display discovered babble instances
+ * and fragments to join / create new groups
  */
 package io.mosaicnetworks.babble.configure;
