+++ babble/src/main/res/layout/fragment_service_discovery.xml
@@ -40,7 +40,7 @@
         android:layout_height="match_parent"
         android:visibility="gone" >
 
-        <io.mosaicnetworks.babble.servicediscovery.mdns.MdnsServicesListView
+        <io.mosaicnetworks.babble.servicediscovery.ServicesListView
             android:id="@+id/servicesListView"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
